Move all calculations for pa01 to their own file
